requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
django==4.2.27
django-celery-results==2.6.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
kombu==5.6.1
packaging==25.0
prompt-toolkit==3.0.52
python-dateutil==2.9.0.post0
redis==7.0.1
six==1.17.0
sqlparse==0.5.5
typing-extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
